requirements.txt
opencv-python
cv2-enumerate-cameras
edge-tts
pyside6
mediapipe
pvporcupine
azure-ai-vision-imageanalysis
pydub